Rename accordion-item
