+++ packages/block-library/src/accordion-content/block.json
@@ -1,11 +1,11 @@
 {
 	"$schema": "https://schemas.wp.org/trunk/block.json",
 	"apiVersion": 3,
-	"name": "core/accordion-item",
+	"name": "core/accordion-content",
 	"version": "0.1.0",
-	"title": "Accordion Item",
+	"title": "Accordion Content",
 	"category": "design",
-	"description": "A single accordion that displays a header and expandable content.",
+	"description": "A section of content within an Accordion that displays a header and expandable content.",
 	"example": {},
 	"__experimental": true,
 	"parent": [ "core/accordion" ],
