+++ packages/block-library/src/accordion-content/index.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Server-side rendering of the `core/accordion-item` block.
+ * Server-side rendering of the `core/accordion-content` block.
  *
  * @package WordPress
  * @since 6.6.0
@@ -11,13 +11,13 @@
  *
  * @return string Returns the updated markup.
  */
-function block_core_accordion_item_render( $attributes, $content ) {
+function block_core_accordion_content_render( $attributes, $content ) {
 	if ( ! $content ) {
 		return $content;
 	}
 
 	$p         = new WP_HTML_Tag_Processor( $content );
-	$unique_id = wp_unique_id( 'accordion-item-' );
+	$unique_id = wp_unique_id( 'accordion-content-' );
 
 	// Initialize the state of the item on the server using a closure,
 	// since we need to get derived state based on the current context.
@@ -31,13 +31,13 @@ function block_core_accordion_item_render( $attributes, $content ) {
 		)
 	);
 
-	if ( $p->next_tag( array( 'class_name' => 'wp-block-accordion-item' ) ) ) {
+	if ( $p->next_tag( array( 'class_name' => 'wp-block-accordion-content' ) ) ) {
 		$open_by_default = $attributes['openByDefault'] ? 'true' : 'false';
 		$p->set_attribute( 'data-wp-context', '{ "id": "' . $unique_id . '", "openByDefault": ' . $open_by_default . ' }' );
 		$p->set_attribute( 'data-wp-class--is-open', 'state.isOpen' );
 		$p->set_attribute( 'data-wp-init', 'callbacks.initIsOpen' );
 
-		if ( $p->next_tag( array( 'class_name' => 'accordion-item__toggle' ) ) ) {
+		if ( $p->next_tag( array( 'class_name' => 'accordion-content__toggle' ) ) ) {
 			$p->set_attribute( 'data-wp-on--click', 'actions.toggle' );
 			$p->set_attribute( 'id', $unique_id );
 			$p->set_attribute( 'aria-controls', $unique_id . '-panel' );
@@ -58,16 +58,16 @@ function block_core_accordion_item_render( $attributes, $content ) {
 }
 
 /**
- * Registers the `core/accordion-item` block on server.
+ * Registers the `core/accordion-content` block on server.
  *
  * @since 6.6.0
  */
-function register_block_core_accordion_item() {
+function register_block_core_accordion_content() {
 	register_block_type_from_metadata(
-		__DIR__ . '/accordion-item',
+		__DIR__ . '/accordion-content',
 		array(
-			'render_callback' => 'block_core_accordion_item_render',
+			'render_callback' => 'block_core_accordion_content_render',
 		)
 	);
 }
-add_action( 'init', 'register_block_core_accordion_item' );
+add_action( 'init', 'register_block_core_accordion_content' );
