+++ packages/block-library/src/accordion/style.scss
@@ -1,18 +1,18 @@
-.wp-block-accordion-item {
+.wp-block-accordion-content {
 	display: grid;
 	grid-template-rows: max-content 0fr;
 }
 
-.wp-block-accordion-item.is-open {
+.wp-block-accordion-content.is-open {
 	grid-template-rows: max-content 1fr;
 }
 
-.accordion-item__heading {
+.accordion-content__heading {
 	margin-block-start: 0;
 	margin-block-end: 0;
 }
 
-.accordion-item__toggle {
+.accordion-content__toggle {
 	font-family: inherit;
 	font-size: inherit;
 	font-weight: inherit;
@@ -34,7 +34,7 @@
 	width: 100%;
 }
 
-.accordion-item__toggle > span {
+.accordion-content__toggle > span {
 	width: 100%;
 }
 
@@ -49,43 +49,43 @@
 }
 
 /* No icon block style */
-.is-style-no-icon .accordion-item__toggle-icon {
+.is-style-no-icon .accordion-content__toggle-icon {
 	background-color: unset;
 }
 
-.wp-block-accordion-header.icon-position-left .accordion-item__toggle {
+.wp-block-accordion-header.icon-position-left .accordion-content__toggle {
 	/* stylelint-disable-next-line declaration-property-value-allowed-list -- This should be refactored to not use the row-reverse value. */
 	flex-direction: row-reverse;
 }
 
 /* RTL language support */
-[dir="rtl"] .wp-block-accordion-header:not(.icon-position-left) .accordion-item__toggle {
+[dir="rtl"] .wp-block-accordion-header:not(.icon-position-left) .accordion-content__toggle {
 	/* stylelint-disable-next-line declaration-property-value-allowed-list -- Automatically adjust icon position for RTL languages. */
 	flex-direction: row-reverse;
 }
 
-[dir="rtl"] .wp-block-accordion-header.icon-position-left .accordion-item__toggle {
+[dir="rtl"] .wp-block-accordion-header.icon-position-left .accordion-content__toggle {
 	flex-direction: row;
 }
 
-.accordion-item__toggle:focus-visible {
+.accordion-content__toggle:focus-visible {
 	outline: 2px solid -webkit-focus-ring-color;
 	outline-offset: 2px;
 }
 
 /* Add transitions only for users who do not prefer reduced motion */
 @media (prefers-reduced-motion: no-preference) {
-	.wp-block-accordion-item .accordion-item__toggle-icon {
+	.wp-block-accordion-content .accordion-content__toggle-icon {
 		transition: transform 0.2s ease-in-out;
 	}
 
-	.wp-block-accordion-item {
+	.wp-block-accordion-content {
 		transition: grid-template-rows 0.3s ease-out;
 	}
 }
 
 .is-open {
-	.accordion-item__toggle-icon.has-icon-plus {
+	.accordion-content__toggle-icon.has-icon-plus {
 		transform: rotate(45deg);
 	}
 }
