+++ packages/block-library/src/index.js
@@ -21,7 +21,7 @@ import {
 //
 // See https://github.com/WordPress/gutenberg/pull/40655 for more context.
 import * as accordion from './accordion';
-import * as accordionItem from './accordion-item';
+import * as accordionContent from './accordion-content';
 import * as accordionHeader from './accordion-header';
 import * as accordionPanel from './accordion-panel';
 import * as archives from './archives';
@@ -241,7 +241,7 @@ const getAllBlocks = () => {
 
 	if ( window?.__experimentalEnableBlockExperiments ) {
 		blocks.push( accordion );
-		blocks.push( accordionItem );
+		blocks.push( accordionContent );
 		blocks.push( accordionHeader );
 		blocks.push( accordionPanel );
 	}
